Big Changes

+++ src/main/java/uz/najottalim/bankingapp/repository/AuthorityRepository.java
@@ -2,13 +2,14 @@ package uz.najottalim.bankingapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import uz.najottalim.bankingapp.model.Authority;
-import uz.najottalim.bankingapp.model.Role;
-
+import uz.najottalim.bankingapp.models.Authority;
+import uz.najottalim.bankingapp.models.Role;
 
 import java.util.List;
 
 @Repository
-public interface AuthorityRepository extends JpaRepository<Authority,Integer> {
+public interface AuthorityRepository extends JpaRepository<Authority, Long> {
     List<Authority> findByRole(Role role);
+
+
 }
